Rebase and Split shard operator integration tests into v1 and v2
