+++ integration-tests/integration-tests-common/src/main/java/com/redhat/service/smartevents/integration/tests/common/TimeUtils.java
@@ -1,4 +1,4 @@
-package com.redhat.service.smartevents.shard.operator.cucumber.common;
+package com.redhat.service.smartevents.integration.tests.common;
 
 import java.time.Duration;
 import java.time.Instant;
