+++ integration-tests/integration-tests-common/src/main/java/com/redhat/service/smartevents/integration/tests/logs/LogConfig.java
@@ -1,4 +1,4 @@
-package com.redhat.service.smartevents.shard.operator.cucumber.logs;
+package com.redhat.service.smartevents.integration.tests.logs;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
