--- integration-tests/shard-operator-integration-tests/shard-operator-integration-tests-common/pom.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-  <parent>
-    <groupId>com.redhat.service.smartevents</groupId>
-    <artifactId>shard-operator-integration-tests</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <relativePath>../pom.xml</relativePath>
-  </parent>
-
-  <modelVersion>4.0.0</modelVersion>
-  <artifactId>shard-operator-integration-tests-common</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <name>SmartEvents :: Shard Operator Integration Tests Common</name>
-
-  <dependencies>
-    <dependency>
-      <groupId>com.redhat.service.smartevents</groupId>
-      <artifactId>shard-operator</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>io.cucumber</groupId>
-      <artifactId>cucumber-java</artifactId>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.fabric8</groupId>
-      <artifactId>openshift-client</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava</artifactId>
-    </dependency>
-  </dependencies>
-
-</project>
