+++ integration-tests/shard-operator-integration-tests/shard-operator-integration-tests-v1/src/test/java/com/redhat/service/smartevents/shard/operator/cucumber/steps/BridgeExecutorSteps.java
@@ -6,9 +6,9 @@ import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.util.concurrent.TimeoutException;
 
+import com.redhat.service.smartevents.integration.tests.common.TimeUtils;
+import com.redhat.service.smartevents.integration.tests.context.ShardContext;
 import com.redhat.service.smartevents.shard.operator.core.utils.LabelsBuilder;
-import com.redhat.service.smartevents.shard.operator.cucumber.common.Context;
-import com.redhat.service.smartevents.shard.operator.cucumber.common.TimeUtils;
 import com.redhat.service.smartevents.shard.operator.v1.resources.BridgeExecutor;
 
 import io.cucumber.java.en.Then;
@@ -22,9 +22,9 @@ import io.fabric8.kubernetes.api.model.SecretBuilder;
  */
 public class BridgeExecutorSteps {
 
-    private Context context;
+    private ShardContext context;
 
-    public BridgeExecutorSteps(Context context) {
+    public BridgeExecutorSteps(ShardContext context) {
         this.context = context;
     }
